# api/openapi/customer-context.yaml
openapi: 3.0.3
info:
  title: Customer Context Service API
  version: 1.0.0
  description: Customer profile and credit management APIs.
  license:
    name: Proprietary
    url: https://openfinance.example.com/license
servers:
  - url: https://api.sandbox.openfinance.ae
    description: Sandbox

tags:
  - name: CustomerManagement

paths:
  /api/v1/customers:
    post:
      tags: [CustomerManagement]
      summary: Create customer
      operationId: createCustomer
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/customers/{customerId}:
    get:
      tags: [CustomerManagement]
      summary: Get customer by id
      operationId: getCustomer
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/customers/{customerId}/credit:
    get:
      tags: [CustomerManagement]
      summary: Get a customer's credit position
      operationId: getCustomerCredit
      description: |-
        Credit limit, used and available credit, without personal data. For
        services that decide on credit (SERVICE realm role, calling client on
        the service's allowed list), staff, and the customer themself.
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/CustomerId'
      responses:
        '200':
          description: Credit position
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerCreditResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/customers/{customerId}/credit-limit:
    put:
      tags: [CustomerManagement]
      summary: Update customer credit limit
      operationId: updateCreditLimit
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/CustomerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCreditLimitRequest'
      responses:
        '200':
          description: Credit limit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/customers/{customerId}/credit/reserve:
    post:
      tags: [CustomerManagement]
      summary: Reserve credit
      operationId: reserveCredit
      description: |-
        Called by other services (SERVICE realm role, calling client on the
        service's allowed list) and staff. x-idempotency-key is required: a
        repeat with the same key and amount returns the current position
        without moving credit again; the same key with a different movement is
        a 409.
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditMovementRequest'
      responses:
        '200':
          description: Credit reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Missing x-idempotency-key or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency key reused for a different credit movement, or a concurrent update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Not enough available credit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/customers/{customerId}/credit/release:
    post:
      tags: [CustomerManagement]
      summary: Release reserved credit
      operationId: releaseCredit
      description: |-
        Called by other services (SERVICE realm role, calling client on the
        service's allowed list) and staff. x-idempotency-key is required: a
        repeat with the same key and amount returns the current position
        without moving credit again; the same key with a different movement is
        a 409.
      security:
        - bearerAuth: []
          dpopAuth: []
      parameters:
        - $ref: '#/components/parameters/Authorization'
        - $ref: '#/components/parameters/DPoP'
        - $ref: '#/components/parameters/InteractionId'
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditMovementRequest'
      responses:
        '200':
          description: Credit released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Missing x-idempotency-key or invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency key reused for a different credit movement, or a concurrent update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    dpopAuth:
      type: http
      scheme: bearer
      bearerFormat: DPoP
  parameters:
    Authorization:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
      description: Access token.
    DPoP:
      name: DPoP
      in: header
      required: true
      schema:
        type: string
      description: DPoP proof JWT.
    InteractionId:
      name: X-FAPI-Interaction-ID
      in: header
      required: true
      schema:
        type: string
      description: End-to-end correlation identifier.
    CustomerId:
      name: customerId
      in: path
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: x-idempotency-key
      in: header
      required: true
      description: |-
        Caller-chosen key, unique per customer, that makes a credit movement
        safe to retry. Derive it from what the movement is for (for example
        "<loanId>:reserve") so a retry resends the same key.
      schema:
        type: string
        maxLength: 128
  schemas:
    CreateCustomerRequest:
      type: object
      required: [firstName, lastName, email, initialCreditLimit]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        initialCreditLimit:
          type: number
          format: double
        currency:
          type: string
          example: USD

    UpdateCreditLimitRequest:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          example: USD

    CreditMovementRequest:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: number
          format: double
        currency:
          type: string
          example: USD
        reference:
          type: string
          maxLength: 128
          description: What the credit is moved for, for example the loan id. Stored with the movement for reconciliation.

    CustomerCreditResponse:
      type: object
      required: [customerId, currency, creditLimit, usedCredit, availableCredit]
      properties:
        customerId:
          type: string
        currency:
          type: string
        creditLimit:
          type: number
          format: double
        usedCredit:
          type: number
          format: double
        availableCredit:
          type: number
          format: double

    CustomerResponse:
      type: object
      required: [customerId, firstName, lastName, creditLimit, usedCredit, availableCredit, status, createdAt, lastModifiedAt]
      properties:
        customerId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
          nullable: true
          description: Absent for customers migrated from the monolith, which never stored an e-mail address.
        phoneNumber:
          type: string
          nullable: true
        creditLimit:
          type: number
          format: double
        usedCredit:
          type: number
          format: double
        availableCredit:
          type: number
          format: double
        status:
          type: string
        creditScore:
          type: integer
          nullable: true
        monthlyIncome:
          type: number
          format: double
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastModifiedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required: [code, message, interactionId, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        interactionId:
          type: string
        timestamp:
          type: string
          format: date-time